requirements.txt
albumentations
matplotlib
numpy
opencv-python
pandas
Pillow
pyparsing
timm
torch
torchvision
tqdm
hydra-core
wandb
